Configure, run-ssh: No need to use ./\$exe
Previous, if run-ssh was passed ./TEST, it would've ended up running
././TEST in the target system.  Similarly, if you passed 'uname',
it would've tried running ./uname, which would've failed
spectacularly.

This commit changes run-ssh so that it won't add the extra './';
this means that shell builtins can be run in the target system.
While this is not generally portable, it's useful in specific
situations, such as in the hints files.